Added ApplyColormap.osl
This script outputs a color depending on the height- and moisturemap that are used as inputs to the script.

Currently the script simply looks up the Colormap and applies color according to the height and moistureness. This means that if the minimumheight is very low, we may color something in watercolor, that shouldn't be water.