The Colormap is now divided into landmass and water
CustomColorMix.osl -> Changed the comment to make the intent a little clearer.

ApplyColormap.osl -> The colormap is now divided into 2 maps, beeing the landmass map and the water map. This allows for depth representation in the water aswell as on the land. Before you could only set the index for the Displacement to the lowest possible value IN HOPES that it represents water.
This works by using the range of the possible Height [0; 1] and break it down into two ranges (for land and water each), beeing the water range [0; MinimumHeight] and the land ]MinimumHeight; 1]. Its now clearly defined, where water is and where land is. For the lookup of the Colormap, we now use the range, the height is in and scale that range (and thus the height) back to the [0; 1] range which allows us to look up the complete water/landmap with a relative height starting at the lowest point of each map.

With this we also solved the problem of higher landmasses not responding to the changing height of the water (which includes that beach sides are typically more green/sandy).